backend/bot/requirements.txt
# Bot dependencies (lightweight)
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
